+++ packages/core/webpack/src/utils/types.ts
@@ -1,0 +1,10 @@
+import {
+  MoveFilesIntoDistFolderPluginOptions,
+  UseClientInjectionPluginOptions,
+} from "../plugins";
+
+export type WebpackBaseConfig = {
+  dtsBundleConfig?: { entries: string[] };
+  packageJsonConfig?: MoveFilesIntoDistFolderPluginOptions;
+  useClientInjectionOptions?: UseClientInjectionPluginOptions;
+};