+++ packages/react/src/atomic/Button/src/index.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { Button } from "./Button";
 
 export { Button } from "./Button";