+++ packages/react/src/atomic/Input/src/index.ts
@@ -1,3 +1,5 @@
+"use client";
+
 import { Input } from "./Input";
 
 export { Input } from "./Input";