+++ packages/react/src/composite/Tabs/src/tabs.stories.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState } from "react";
 import type { Meta, StoryObj } from "@storybook/react";
 import { Box } from "@nimbus-ds/box";
 import { Text } from "@nimbus-ds/text";
@@ -66,5 +66,53 @@ export const basic: Story = {
 export const fullWidth: Story = {
   args: {
     fullWidth: true,
+  },
+};
+
+export const controlled: Story = {
+  render: () => {
+    const [selectedTab, setSelectedTab] = useState(0);
+    const handleSelect = (index: number) => setSelectedTab(index);
+    
+    return (
+      <Tabs selected={selectedTab} onTabSelect={handleSelect}>
+        <Tabs.Item label="Tab 1" labelContent={<Tag>Controlled</Tag>}>
+          <Box
+            borderColor="neutral-interactive"
+            borderStyle="dashed"
+            borderWidth="1"
+            padding="2"
+          >
+            <Text fontSize="base" textAlign="center">
+              Replace me with your controlled tab 1 content
+            </Text>
+          </Box>
+        </Tabs.Item>
+        <Tabs.Item label="Tab 2">
+          <Box
+            borderColor="neutral-interactive"
+            borderStyle="dashed"
+            borderWidth="1"
+            padding="2"
+          >
+            <Text fontSize="base" textAlign="center">
+              Replace me with your controlled tab 2 content
+            </Text>
+          </Box>
+        </Tabs.Item>
+        <Tabs.Item label="Tab 3">
+          <Box
+            borderColor="neutral-interactive"
+            borderStyle="dashed"
+            borderWidth="1"
+            padding="2"
+          >
+            <Text fontSize="base" textAlign="center">
+              Replace me with your controlled tab 3 content
+            </Text>
+          </Box>
+        </Tabs.Item>
+      </Tabs>
+    );
   },
 };