+++ packages/react/src/atomic/Skeleton/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The skeleton is a basic component that displays a placeholder preview of its content before the data is loaded, reducing the feeling of slow loading time.
 
+## 2025-03-18 `2.2.1`
+
+### 💡 Others
+
+- Rebuild after build process changes to add support for modular imports and Server Components. ([#276](https://github.com/TiendaNube/nimbus-design-system/pull/276) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2023-03-13 `2.2.0`
 
 ### 💡 Others