+++ packages/react/src/atomic/Spinner/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The Spinner component is used to inform the processing of tasks.
 
+## 2025-03-18 `3.1.0`
+
+#### 🎉 New features
+
+- Adds `use-client` directive to the build output in order to support NextJS applications. ([#276](https://github.com/TiendaNube/nimbus-design-system/pull/276) by [@joacotornello](https://github.com/joacotornello))
+
+### 💡 Others
+
+- Rebuild after build process changes to add support for modular imports and Server Components. ([#276](https://github.com/TiendaNube/nimbus-design-system/pull/276) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2023-03-13 `3.0.0`
 
 ### 🛠 Breaking changes