+++ packages/react/src/atomic/Text/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 Text is a basic component that allows us to write blocks of text and give it formatting to use within other components, sections and pages of our application or website.
 
+## 2025-03-25 `6.4.0`
+
+### 🎉 New features
+
+- Added `textDecoration` property to `Text` component. ([#279](https://github.com/TiendaNube/nimbus-design-system/pull/279) by [@guidobotta](https://github.com/guidobotta))
+
+## 2025-03-18 `6.3.1`
+
+### 💡 Others
+
+- Rebuild after build process changes to add support for modular imports and Server Components. ([#276](https://github.com/TiendaNube/nimbus-design-system/pull/276) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2023-06-13 `6.3.0`
 
 #### 🎉 New features